use podman unshare to change ownership + bump quay version
When podman runs containers in rootless mode, it remaps the container's
UIDs to the host UIDs assigned to the user. For example, the container's
UID 1001 is no longer 1001 on the host; it will be remapped to something
like 101000. Therefore, we need to use podman unshare to assign the
correct owner UID to the directory.