db: Replace Postgres database with an in-app Sqlite db for Quay during fresh install of OMR (#152)
* Remove postgres references from ansible tasks

Signed-off-by: harishsurf <hgovinda@redhat.com>

* Update ansible tasks to install quay with sqlite

* Clean up remaining postgres references

* Add debug logs when quay-app systemd service fails

* change dir path for sqlite db file to live outside of /quay-registry path

* set WORKER_COUNT=1 when starting quay with Sqlite

without this, multiple workers tend to write simultaneously
causing concurrent write issues

* Add conditional when creating volume based on user input

* use podman unshare to change ownership + bump quay version

When podman runs containers in rootless mode, it remaps the container's
UIDs to the host UIDs assigned to the user. For example, the container's
UID 1001 is no longer 1001 on the host; it will be remapped to something
like 101000. Therefore, we need to use podman unshare to assign the
correct owner UID to the directory.

---------

Signed-off-by: harishsurf <hgovinda@redhat.com>